+++ test/test_netconf_monitoring.sh
@@ -171,9 +171,12 @@ EOF
         continue
     fi
     # Mask netconf header and footer
-    sed -i -e "s/<rpc-reply $DEFAULTNS><data xmlns=\"urn:ietf:params:xml:ns:yang:ietf-netconf-monitoring\">//" -e 's/<\/data><\/rpc-reply>]]>]]>//' /var/tmp/test_netconf_monitoring.sh/ex.yang
+    sed -i -e "s/<rpc-reply $DEFAULTNS><data xmlns=\"urn:ietf:params:xml:ns:yang:ietf-netconf-monitoring\">//" /var/tmp/test_netconf_monitoring.sh/ex.yang
+    sed -i -e 's/<\/data><\/rpc-reply>]]>]]>//' /var/tmp/test_netconf_monitoring.sh/ex.yang
     # Decode XML
-    sed -i -e 's/&gt;/>/g' -e 's/&lt;/</g' -e 's/\&amp;/\&/g' /var/tmp/test_netconf_monitoring.sh/ex.yang
+    sed -i -e 's/&gt;/>/g' /var/tmp/test_netconf_monitoring.sh/ex.yang
+    sed -i -e 's/&lt;/</g' /var/tmp/test_netconf_monitoring.sh/ex.yang
+    sed -i -e 's/\&amp;/\&/g' /var/tmp/test_netconf_monitoring.sh/ex.yang
     new "get-schema check yang $b"
     diff $dir/ex.yang $f
     if [ $? -ne 0 ]; then