+++ CHANGELOG.md
@@ -72,6 +72,7 @@ Developers may need to change their code
     * Added netconf ssh subsystem
     * Renamed from `clixon` built in `docker/base`
 * C-API
+  * Changed type of `veclen` parameter to `size_t` in `xpath_vec_flag()`
   * Added `with-defaults` parameter (default 0) to `xmldb_get0()`
   * Added `sock_flags` parameter to `clixon_proc_socket()`